src/test/java/racingcar/model/CarTest.java
package racingcar.model;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;
import org.junit.jupiter.params.provider.ValueSource;

import static org.assertj.core.api.Assertions.*;

public class CarTest {

    @ParameterizedTest
    @ValueSource(strings = {"123456", ""})
    void 자동차_이름은_1자_이상_5자_이하(String input) {
        assertThatThrownBy(() -> new Car(input)).isInstanceOf(IllegalArgumentException.class);
    }

    @ParameterizedTest
    @CsvSource(value = {"1:0", "2:0", "3:0", "4:1", "5:1", "6:1", "7:1", "8:1", "9:1"}, delimiter = ':')
    void random값이_4이상이면_전진(int input, int expected) {
        Car car = new Car("hello");
        car.move(input);
        assertThat(car.getDistance()).isEqualTo(expected);
    }
}
